upd, bot creates two batch orders
