+++ README.md
@@ -1,0 +1,11 @@
+## Grid bot for Deribit and dydx
+
+### Functionality:
+- Grid trading with limit orders one instrument
+- Grid trading with limit orders two instruments. In other words, the bot imitates trading behavior as if such a pair of two exotic instruments exists on the exchange.
+
+
+Run tests with pytest:
+```sh
+    python3 -m pytest -s
+```
