+++ constants/constants.py
@@ -4,11 +4,11 @@ from utils import add_get_attributes
 
 
 NDIGITS_PRICES_ROUNDING = {
-    DeribitPerpetualMarkets.BTC_PERPETUAL: 2,
-    DeribitPerpetualMarkets.ETH_PERPETUAL: 2,
+    DeribitPerpetualMarkets.BTC_PERPETUAL: 0,
+    DeribitPerpetualMarkets.ETH_PERPETUAL: 1,
     DeribitSpotMarkets.ETH_BTC: 4,
-    DeribitSpotMarkets.ETH_USDC: 4,
-    DeribitSpotMarkets.BTC_USDC: 4
+    DeribitSpotMarkets.ETH_USDC: 2,
+    DeribitSpotMarkets.BTC_USDC: 2
 }
 
 NDIGITS_AMOUNTS_ROUNDING = {
