+++ utils/__init__.py
@@ -3,3 +3,4 @@ from .config_setup import *
 from .error_checkers import *
 from .sides_directions import *
 from .spread_calculation import *
+from .trade_utils import *
